clean up.

+++ contracts/converter/interfaces/IDFEngine.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.5.2;
 
 contract IDFEngine {
-    function deposit(address _sender, address _tokenID, uint256 _amount) public returns (uint256);
-    function withdraw(address _sender, address _tokenID, uint256 _amount) public returns (uint256);
-    function destroy(address _sender, uint256 _amount) public returns (bool);
+    function deposit(address _sender, address _tokenID, uint _amount) public returns (uint);
+    function withdraw(address _sender, address _tokenID, uint _amount) public returns (uint);
+    function destroy(address _sender, uint _amount) public returns (bool);
 }
