+++ contracts/token/interfaces/IDSToken.sol
@@ -3,6 +3,6 @@ pragma solidity ^0.5.2;
 import './IERC20Token.sol';
 
 contract IDSToken is IERC20Token {
-    function mint(address _account, uint256 _value) public;
-    function burn(address _account, uint256 _value) public;
+    function mint(address _account, uint _value) public;
+    function burn(address _account, uint _value) public;
 }
