+++ contracts/token/interfaces/IERC20Token.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.5.2;
     ERC20 Standard Token interface
 */
 contract IERC20Token {
-    function balanceOf(address _owner) public view returns (uint256);
-    function allowance(address _owner, address _spender) public view returns (uint256);
-    function transfer(address _to, uint256 _value) public returns (bool success);
-    function transferFrom(address _from, address _to, uint256 _value) public returns (bool success);
-    function approve(address _spender, uint256 _value) public returns (bool success);
+    function balanceOf(address _owner) public view returns (uint);
+    function allowance(address _owner, address _spender) public view returns (uint);
+    function transfer(address _to, uint _value) public returns (bool success);
+    function transferFrom(address _from, address _to, uint _value) public returns (bool success);
+    function approve(address _spender, uint _value) public returns (bool success);
 }
